Proposing a Community Discussions Page
Per @ErinBecker "I think it would be nice to have. Maybe it doesn’t need to have it’s own menu pull down item, but could be linked to from the “connect” page?"

The purpose of this page is to provide archived and upcoming community discussions. I'd like to format it similar to https://ropensci.org/commcalls/