Clarify Supporting Instructor description
Thanks for the excellent FAQ!

I find the current description confusing. It currently sounds like a Supporting Instructor will support an instructor with fewer than 3 workshops. The question above indicates it is meant as a role for instructors with fewer than 3 workshops. If my understanding of the role is correct I hope this PR clarifies it.